examples/flux2-klein-9b/pyproject.toml
[project]
name = "flux2-klein-9b"
version = "0.1.0"
description = "Flux2-klein-9B endpoint (base+turbo; inference-only; Cozy manifest via endpoint.toml)"
requires-python = ">=3.12"
dependencies = [
    "gen-worker[torch]>=0.4.3",
    "diffusers @ git+https://github.com/huggingface/diffusers.git@99e2cfff27dec514a43e260e885c5e6eca038b36",
    "transformers>=4.56,<5",
    "accelerate",
    "pillow",
    # Needed for fp8/nvfp4 variants (diffusers TorchAoConfig and/or variant-aware loading).
    "torchao",
]

[dependency-groups]
dev = ["mypy>=1.10.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flux2_klein_9b"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = true
